+++ src/modules/scenarioeditor/ScenarioFiltersTabPanel.vue
@@ -1,0 +1,373 @@
+<script setup lang="ts">
+import PanelHeading from "@/components/PanelHeading.vue";
+import { injectStrict, sortBy } from "@/utils";
+import { activeScenarioKey } from "@/components/injects";
+import { computed, ref, watchEffect } from "vue";
+import { Sidc } from "@/symbology/sidc";
+import { useSelectedItems } from "@/stores/selectedStore";
+import { echelonValues, symbolSetValues } from "@/symbology/values";
+import BaseButton from "@/components/BaseButton.vue";
+import { NUnit } from "@/types/internalModels";
+import { app6d as sym } from "@/symbology/standards/app6d";
+import AccordionPanel from "@/components/AccordionPanel.vue";
+import FilterTree, { NestedUnitStatItem } from "@/modules/scenarioeditor/FilterTree.vue";
+import { IconExpandAllOutline } from "@iconify-prerendered/vue-mdi";
+import IconButton from "@/components/IconButton.vue";
+
+const {
+  store: { state },
+} = injectStrict(activeScenarioKey);
+
+const sideTree = ref<NestedUnitStatItem[]>([]);
+const emtTree = ref<NestedUnitStatItem[]>([]);
+const iconTree = ref<NestedUnitStatItem[]>([]);
+const { selectedUnitIds } = useSelectedItems();
+
+const excludedKeys = ref<Set<string>>(new Set());
+const expandedKeys = ref<string[]>([]);
+const flatStats = ref<Record<string, number>>({});
+
+watchEffect(() => {
+  const stats: Record<string, number> = {};
+  const sideStatItems: NestedUnitStatItem[] = [];
+  const emtStatItems: NestedUnitStatItem[] = [];
+  const iconStatItems: NestedUnitStatItem[] = [];
+  Object.values(state.unitMap).forEach((unit) => {
+    const { sideKey, sideGroupKey, emtKey, symbolSetKey, entityKey, entityTypeKey } =
+      updateUnitStats(unit, stats);
+
+    const iconSidc = new Sidc(unit.sidc);
+    const originalEmt = iconSidc.emt;
+    iconSidc.emt = "00";
+    iconSidc.hqtfd = "0";
+    iconSidc.standardIdentity = "3";
+    iconSidc.entitySubType = "00";
+    iconSidc.modifierOne = "00";
+    iconSidc.modifierTwo = "00";
+    const sidc = iconSidc.toString();
+    iconSidc.mainIcon = "000000";
+    const sidcSymbolSet = iconSidc.toString();
+    iconSidc.emt = originalEmt;
+    const sidcEmt = iconSidc.toString();
+    if (stats[sideKey] === 1) {
+      sideStatItems.push({
+        key: sideKey,
+        label: getSideLabel(unit._sid),
+        sidc: sidc,
+      });
+    }
+    if (stats[sideGroupKey] === 1) {
+      const sideItem = sideStatItems.find((item) => item.key === sideKey);
+      if (sideItem) {
+        const children = sideItem.children || [];
+        children.push({
+          key: sideGroupKey,
+          label: getSideGroupLabel(unit._gid),
+          sidc: "10031000100000000000",
+        });
+
+        sideItem.children = children;
+        sideItem.sidc = "10031000100000000000";
+      }
+    }
+    if (stats[emtKey] === 1) {
+      emtStatItems.push({ key: emtKey, label: getEchelonLabel(emtKey), sidc: sidcEmt });
+    }
+    if (stats[symbolSetKey] === 1) {
+      iconStatItems.push({
+        key: symbolSetKey,
+        label: getIconLabel({ symbolSet: symbolSetKey }),
+        sidc: sidcSymbolSet,
+      });
+    }
+    if (stats[entityKey] === 1) {
+      const iconItem = iconStatItems.find((item) => item.key === symbolSetKey);
+      if (iconItem) {
+        const children = iconItem.children || [];
+        children.push({
+          key: entityKey,
+          label: getIconLabel({
+            symbolSet: symbolSetKey,
+            entity: entityKey.split("-")[1],
+          }),
+          sidc: sidc,
+        });
+        iconItem.children = sortBy(children, "label");
+      }
+    }
+    if (stats[entityTypeKey] === 1) {
+      const entityItem = iconStatItems
+        .find((item) => item.key === symbolSetKey)
+        ?.children?.find((item) => item.key === entityKey);
+      if (entityItem) {
+        const children = entityItem.children || [];
+        children.push({
+          key: entityTypeKey,
+          label: getIconLabel({
+            symbolSet: symbolSetKey,
+            entity: entityKey.split("-")[1],
+            entityType: entityTypeKey.split("-")[2],
+          }),
+          sidc: sidc,
+        });
+        entityItem.children = sortBy(children, "label");
+      }
+    }
+  });
+
+  flatStats.value = stats;
+  sideTree.value = sortBy(sideStatItems, "label");
+  emtTree.value = sortBy(emtStatItems, "label");
+  iconTree.value = sortBy(iconStatItems, "label");
+});
+
+const selectedStats = computed(() => {
+  const stats: Record<string, number> = {};
+  selectedUnitIds.value.forEach((unitId) => {
+    updateUnitStats(unitId, stats);
+  });
+  return stats;
+});
+
+function updateUnitStats(unitOrUnitId: string | NUnit, stats: Record<string, number>) {
+  const unit =
+    typeof unitOrUnitId === "string" ? state.unitMap[unitOrUnitId] : unitOrUnitId;
+  const keys = createKeys(unit);
+  const { symbolSetKey, entityKey, entityTypeKey, emtKey, sideKey, sideGroupKey } = keys;
+  stats[symbolSetKey] = (stats[symbolSetKey] || 0) + 1;
+  stats[entityKey] = (stats[entityKey] || 0) + 1;
+  stats[entityTypeKey] = (stats[entityTypeKey] || 0) + 1;
+  stats[emtKey] = (stats[emtKey] || 0) + 1;
+  stats[sideKey] = (stats[sideKey] || 0) + 1;
+  stats[sideGroupKey] = (stats[sideGroupKey] || 0) + 1;
+  return keys;
+}
+
+function createKeys(unit: NUnit) {
+  const sidc = new Sidc(unit.sidc);
+  const symbolSetKey = `${sidc.symbolSet}`;
+  const entityKey = `${sidc.symbolSet}-${sidc.entity}`;
+  const entityTypeKey = `${sidc.symbolSet}-${sidc.entity}-${sidc.entityType}`;
+  const emtKey = `emt-${sidc.emt}`;
+  const sideKey = `side-${unit._sid}`;
+  const sideGroupKey = `side-${unit._sid}-${unit._gid}`;
+  return { symbolSetKey, entityKey, entityTypeKey, emtKey, sideKey, sideGroupKey };
+}
+
+function selectByKeys(keys: string[]) {
+  Object.values(state.unitMap).forEach((unit) => {
+    const { symbolSetKey, entityKey, entityTypeKey, emtKey, sideKey, sideGroupKey } =
+      createKeys(unit);
+    if (
+      excludedKeys.value.has(symbolSetKey) ||
+      excludedKeys.value.has(entityKey) ||
+      excludedKeys.value.has(entityTypeKey) ||
+      excludedKeys.value.has(emtKey) ||
+      excludedKeys.value.has(sideKey) ||
+      excludedKeys.value.has(sideGroupKey)
+    ) {
+      return;
+    }
+    if (keys.includes(symbolSetKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+    if (keys.includes(entityKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+    if (keys.includes(entityTypeKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+    if (keys.includes(emtKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+    if (keys.includes(sideKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+    if (keys.includes(sideGroupKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+  });
+}
+
+function selectByKey(key: string) {
+  Object.values(state.unitMap).forEach((unit) => {
+    const { symbolSetKey, entityKey, entityTypeKey, emtKey, sideKey, sideGroupKey } =
+      createKeys(unit);
+
+    if (
+      excludedKeys.value.has(symbolSetKey) ||
+      excludedKeys.value.has(entityKey) ||
+      excludedKeys.value.has(entityTypeKey) ||
+      excludedKeys.value.has(emtKey) ||
+      excludedKeys.value.has(sideKey) ||
+      excludedKeys.value.has(sideGroupKey)
+    ) {
+      return;
+    }
+    if (key === symbolSetKey) {
+      selectedUnitIds.value.add(unit.id);
+    } else if (key === entityKey) {
+      selectedUnitIds.value.add(unit.id);
+    } else if (key === entityTypeKey) {
+      selectedUnitIds.value.add(unit.id);
+    } else if (key === emtKey) {
+      selectedUnitIds.value.add(unit.id);
+    } else if (key === sideKey) {
+      selectedUnitIds.value.add(unit.id);
+    } else if (key === sideGroupKey && !excludedKeys.value.has(sideGroupKey)) {
+      selectedUnitIds.value.add(unit.id);
+    }
+  });
+}
+
+function clearByKey(key: string) {
+  selectedUnitIds.value.forEach((unitId) => {
+    const unit = state.unitMap[unitId];
+    const { symbolSetKey, entityKey, entityTypeKey, emtKey, sideKey, sideGroupKey } =
+      createKeys(unit);
+    if (key === symbolSetKey) {
+      selectedUnitIds.value.delete(unit.id);
+    } else if (key === entityKey) {
+      selectedUnitIds.value.delete(unit.id);
+    } else if (key === entityTypeKey) {
+      selectedUnitIds.value.delete(unit.id);
+    } else if (key === emtKey) {
+      selectedUnitIds.value.delete(unit.id);
+    } else if (key === sideKey) {
+      selectedUnitIds.value.delete(unit.id);
+    } else if (key === sideGroupKey) {
+      selectedUnitIds.value.delete(unit.id);
+    }
+  });
+}
+
+function getEchelonLabel(echelon: string) {
+  if (echelon.startsWith("emt-")) {
+    const [, nechelon] = echelon.split("-");
+    return echelonValues.find((v) => v.code === nechelon)?.text || echelon;
+  }
+  return echelonValues.find((v) => v.code === echelon)?.text || echelon;
+}
+
+function getSymbolSetLabel(symbolSet: string) {
+  return symbolSetValues.find((v) => v.code === symbolSet)?.text || symbolSet;
+}
+
+type IconData = {
+  symbolSet?: string;
+  entity?: string;
+  entityType?: string;
+};
+
+function getSideLabel(sideId: string) {
+  return state.sideMap[sideId]?.name || sideId;
+}
+
+function getSideGroupLabel(sideGroupId: string) {
+  return state.sideGroupMap[sideGroupId]?.name || sideGroupId;
+}
+
+function getIconLabel({ symbolSet, entity, entityType }: IconData) {
+  let label = symbolSet ?? entity ?? entityType ?? "Unknown";
+  if (symbolSet === undefined) {
+    return label;
+  }
+  if (entity == undefined && entityType == undefined) {
+    return getSymbolSetLabel(symbolSet);
+  }
+  if (entity !== undefined && entityType == undefined) {
+    const mainIcon = sym[symbolSet]?.mainIcon;
+    const i = mainIcon?.find((icon) => icon.code.startsWith(entity));
+    return i?.entity || label;
+  }
+  if (entity !== undefined && entityType !== undefined) {
+    const mainIcon = sym[symbolSet]?.mainIcon;
+    const prefix = entity + entityType;
+    const i = mainIcon?.find((icon) => icon.code.startsWith(prefix));
+    return i?.entityType || i?.entity || label;
+  }
+  return label;
+}
+
+function onSelect(event: any) {
+  const key = event.detail.value.key as string;
+
+  if (selectedStats.value[key]) {
+    clearByKey(key);
+  } else {
+    selectByKey(key);
+  }
+}
+
+function expandAllIcons() {
+  const keys = Object.keys(flatStats.value).filter((key) => !key.startsWith("side-"));
+  const expandedSideKeys = keys.filter((key) => key.startsWith("side-"));
+  if (expandedKeys.value.length - expandedSideKeys.length === keys.length) {
+    expandedKeys.value = [];
+  } else {
+    expandedKeys.value = keys;
+  }
+}
+</script>
+<template>
+  <div class="px-4">
+    <header
+      class="sticky top-0 z-10 -mx-4 flex h-12 items-center justify-between bg-white px-4 py-2 dark:bg-gray-800"
+    >
+      <PanelHeading>Select units</PanelHeading>
+      <div class="flex items-center space-x-1">
+        <BaseButton v-if="excludedKeys.size" small @click="excludedKeys.clear()"
+          >Clear excluded
+          <span class="muted-badge">{{ excludedKeys.size }}</span></BaseButton
+        >
+        <BaseButton v-if="selectedUnitIds.size" small @click="selectedUnitIds.clear()"
+          >Clear selected
+          <span class="muted-badge ml-1">{{ selectedUnitIds.size }}</span></BaseButton
+        >
+        <IconButton title="Expand all" @click="expandAllIcons()"
+          ><IconExpandAllOutline
+        /></IconButton>
+      </div>
+    </header>
+    <AccordionPanel label="Unit symbol" defaultOpen>
+      <FilterTree
+        :tree="iconTree"
+        v-model:expandedKeys="expandedKeys"
+        :stats="flatStats"
+        :selectedStats="selectedStats"
+        :excludedKeys="excludedKeys"
+        @select="onSelect"
+        @clear="clearByKey"
+        @exclude="excludedKeys.add($event)"
+        @clearExclude="excludedKeys.delete($event)"
+      />
+    </AccordionPanel>
+
+    <AccordionPanel label="Command level">
+      <FilterTree
+        :tree="emtTree"
+        v-model:expandedKeys="expandedKeys"
+        :stats="flatStats"
+        :selectedStats="selectedStats"
+        :excludedKeys="excludedKeys"
+        @select="onSelect"
+        @clear="clearByKey"
+        @exclude="excludedKeys.add($event)"
+        @clearExclude="excludedKeys.delete($event)"
+      />
+    </AccordionPanel>
+    <AccordionPanel label="Side">
+      <FilterTree
+        :tree="sideTree"
+        v-model:expandedKeys="expandedKeys"
+        :stats="flatStats"
+        :selectedStats="selectedStats"
+        :excludedKeys="excludedKeys"
+        @select="onSelect"
+        @clear="clearByKey"
+        @exclude="excludedKeys.add($event)"
+        @clearExclude="excludedKeys.delete($event)"
+      />
+    </AccordionPanel>
+  </div>
+</template>